users_microservice/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
Markdown==3.4.4
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
